requirements.txt
appdirs==1.4.4
black==20.8b1
click==7.1.2
mypy-extensions==0.4.3
OdooRPC==0.7.0
pathspec==0.8.0
pkg-resources==0.0.0
pycups==2.0.1
regex==2020.9.27
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
